+++ app/src/main/java/com/gitficko/github/ui/home/HomeViewModel.kt
@@ -1,13 +1,87 @@
 package com.gitficko.github.ui.home
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
+import android.app.Application
+import android.content.Intent
+import androidx.browser.customtabs.CustomTabsIntent
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.viewModelScope
+import com.gitficko.github.R
+import com.gitficko.github.model.RemoteGithubUser
+import com.gitficko.github.model.auth.AuthRepository
+import com.gitficko.github.remote.GitHubRep
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.trySendBlocking
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.receiveAsFlow
+import kotlinx.coroutines.launch
+import net.openid.appauth.AuthorizationService
 
-class HomeViewModel : ViewModel() {
+class HomeViewModel(application: Application) : AndroidViewModel(application) {
 
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+    private val authService: AuthorizationService = AuthorizationService(getApplication())
+
+    private val authRepository = AuthRepository()
+    private val userRepository = GitHubRep()
+
+    private val loadingMutableStateFlow = MutableStateFlow(false)
+    private val userInfoMutableStateFlow = MutableStateFlow<RemoteGithubUser?>(null)
+    private val toastEventChannel = Channel<Int>(Channel.BUFFERED)
+    private val logoutPageEventChannel = Channel<Intent>(Channel.BUFFERED)
+    private val logoutCompletedEventChannel = Channel<Unit>(Channel.BUFFERED)
+
+
+    val loadingFlow: Flow<Boolean>
+        get() = loadingMutableStateFlow.asStateFlow()
+
+    val userInfoFlow: Flow<RemoteGithubUser?>
+        get() = userInfoMutableStateFlow.asStateFlow()
+
+    val toastFlow: Flow<Int>
+        get() = toastEventChannel.receiveAsFlow()
+
+    val logoutPageFlow: Flow<Intent>
+        get() = logoutPageEventChannel.receiveAsFlow()
+
+    val logoutCompletedFlow: Flow<Unit>
+        get() = logoutCompletedEventChannel.receiveAsFlow()
+
+
+    fun loadUserInfo() {
+        viewModelScope.launch {
+            loadingMutableStateFlow.value = true
+            runCatching {
+                userRepository.getUserInformation()
+            }.onSuccess {
+                userInfoMutableStateFlow.value = it
+                loadingMutableStateFlow.value = false
+            }.onFailure {
+                loadingMutableStateFlow.value = false
+                userInfoMutableStateFlow.value = null
+                toastEventChannel.trySendBlocking(R.string.get_user_error)
+            }
+        }
     }
-    val text: LiveData<String> = _text
+
+    fun logout() {
+        val customTabsIntent = CustomTabsIntent.Builder().build()
+
+        val logoutPageIntent = authService.getEndSessionRequestIntent(
+            authRepository.getEndSessionRequest(),
+            customTabsIntent
+        )
+
+        logoutPageEventChannel.trySendBlocking(logoutPageIntent)
+    }
+
+    fun webLogoutComplete() {
+        authRepository.logout()
+        logoutCompletedEventChannel.trySendBlocking(Unit)
+    }
+
+    override fun onCleared() {
+        super.onCleared()
+        authService.dispose()
+    }
 }
