--- app/src/main/java/com/gitficko/github/ui/user_info/UserInfoFragment.kt
@@ -1,65 +1,0 @@
-package com.gitficko.github.ui.user_info
-
-import android.app.Activity
-import android.os.Bundle
-import android.view.View
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.view.isVisible
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import by.kirich1409.viewbindingdelegate.viewBinding
-import com.gitficko.github.R
-import com.gitficko.github.databinding.FragmentUserInfoBinding
-import com.gitficko.github.utils.launchAndCollectIn
-import com.gitficko.github.utils.resetNavGraph
-import com.gitficko.github.utils.toast
-
-class UserInfoFragment : Fragment(R.layout.fragment_user_info) {
-
-    private val viewModel: UserInfoViewModel by viewModels()
-    private val binding by viewBinding(FragmentUserInfoBinding::bind)
-
-    private val logoutResponse = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        if (result.resultCode == Activity.RESULT_OK) {
-            viewModel.webLogoutComplete()
-        } else {
-            viewModel.webLogoutComplete()
-        }
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.getUserInfo.setOnClickListener {
-            findNavController().navigate(R.id.action_repositoryListFragment_to_navigation_repositories)
-        }
-        binding.logout.setOnClickListener {
-            viewModel.logout()
-        }
-
-        viewModel.loadingFlow.launchAndCollectIn(viewLifecycleOwner) { isLoading ->
-            binding.progressBar.isVisible = isLoading
-            binding.getUserInfo.isEnabled = !isLoading
-            binding.userInfo.isVisible = !isLoading
-        }
-
-        viewModel.userInfoFlow.launchAndCollectIn(viewLifecycleOwner) { userInfo ->
-            binding.userInfo.text = userInfo?.login
-        }
-
-        viewModel.toastFlow.launchAndCollectIn(viewLifecycleOwner) {
-            toast(it)
-        }
-
-        viewModel.logoutPageFlow.launchAndCollectIn(viewLifecycleOwner) {
-            logoutResponse.launch(it)
-        }
-
-        viewModel.logoutCompletedFlow.launchAndCollectIn(viewLifecycleOwner) {
-            findNavController().resetNavGraph(R.navigation.nav_graph)
-        }
-    }
-}
